Added a script to get the latest build version-revision with option deployment and architecture options.
Modified the build-base-docker-images script to query the build config DB to get the latest
version-revision if not specified.

Modifed the build-base-image script to stage the rabbitmq key to the image so that the
install_dependencies.sh script doesn't have to use wget.

I couldn't test completely because of network issues.